refactor: moved enumerable extentions to own file
